separate runZ3Web from runZ3File
